pyproject.toml
[project]
name = "db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "yfinance>=0.2.66",
]
